changed the way we access copy number
IMP.atom.get_copy_index will traverse up the Hierarchy to get the copy number for a bead. 
Copy number is always set up for a molecule if created using PMI: create_molecule seems to use a default constructor with copy number = 0. So the committed code seems a tad cleaner. Please revert/complain if this does not work for you.