Add summary JSON metadata and link from RMF
Write all of the information on the clustering
to a machine-readable JSON file, and add a link
to that file from the generated cluster centroid
RMF. This should allow us to extract all of the
information on both the modeling and the clustering
from the RMF, in order to populate the final mmCIF
file. Closes #38.